Fix issue with `lfe_eval:expr/2`
This adds a case clause in `lfe_eval:eval_apply/3`, which fixes an
issue with `lfe_eval:expr/2`. The case clause is needed for calling
functions bound to the environment passed to `lfe_eval:expr/2`.
Without that case clause, calling a function bound to that very
environment causes a case clause error in `lfe_eval:eval_apply/3`.