Added FizzBuzz example
Added a fizzbuzz solution to the examples folder. I tried to use
a range of the different features within LFE/Erlang so it wasn't
a plane and boring example. There is pattern matching, guards,
higher order functions, and tail recursion(?).

The recursive attempt (fizzbuzz:buzz2/1) gives some interesting
results when printing in the repl. The results seem to be correct
in that the individual results are correct. But displaying them
seems to be pretty entertaining.