Added support for #| foo |# style block comments.
This adds support for the following comment style

(defun (foo a b)
       (+ a b #| foo |# 3))

This is also legal

(defun foo
       #| (let ((x (: io read_file 'bar)))
               (do-stuff x)) |#
               (no-really-do-stuff (: gief 'bar)))

As has been covered in every mention of this topic, this is not legal
and never will be under the current implementation of the parser.

(defun foo
       #| (let (#| get info |# (x (: io read_file 'bar)))
               (do-stuff x)) |#
               (+ 3 3))

Nested comment blocks are not allowed and will error as a syntax error.