Add nicer parse error for nested comments.
The parser does not support nested block comments and the debate rages
on as to whether or not they are required. This is an improvement on the
existing support for simple block comments in that it provides a syntax
error and a line number when they are encountered.

I'm not convinced I've done it the right way but it's a state. :)