c_src/lfeexec.c: Correctly size the buffer used for the path name.
The needed size depends on user input (it's taken from an environment
variable) so using a fixed size would only work if environment
variable had an upper limit on size.  It's much easier to just
allocate the buffer as needed to avoid any chance of an buffer
overflow via this vector.