Fix #183
Add clause to lfe_io_pretty:term/4 to handle (map ...) forms as having
map bodies, abstracting away F(unc) and P(refix) in map/4 to map/6,
for maximum code reuse.

Properly indent both #m(...) and (map ...) forms when prettifying.

Perform case analysis on the number of elements in the tail of a
(map ...) form and ensure inclusion of the last element, even if it
makes for an odd number.

Update map/4 documentation to more accurately reflect its behaviour.