Update Makefile
Allow DESTBINDIR to be set by the user.

I use Nix to manage Erlang installations, which results in my not being
able to use the default DESTBINDIR.

With this change, the following works for me in fish shell:

    env DESTBINDIR=/usr/local/bin make install

This shouldn't have any negative consequences for "normal" users and
might be helpful to others as well.