Move visual "tests to dev and update .travis.yml
The visual "tests" are not meant to be compiled or run as such and
break the build on older OTP releases on Travis. Per discussion, move
them to dev so Travis won't try to build them.

Re-enable R16B03-1, R15B03 and R14B04 to see if they succeed now.