Add implicit equality tests of recurring variables in patterns
We add an implicit equality test for recurring variables in
patterns. Variable scoping is still observed so explicit tests need
to be added for variables already bound before the match. This differs
from Erlang but we have kept importing variables into bitseg lengths.

This is the first test version and will at least need cleaning up.