Use stqndard erlang handling of records
This is necessary as now guards are the same as Erlag guards and we
cannot bind variables. This means the old record expansions are no
longer usable. This is still needs some more work to get the right LFE
record primitives. Can't yet use records in the shell but we are
getting there.