Removed specific handling for match specification for records
These were somee extra features the were trying to be helpful and
automatically added _='_' options to handling for use in match
specifications. This was done with a 'emp-<record>' macro and special
handling in lfe_ms.erl for records in patterns. This is easy and
straightforward to do and could cause confusion end errors if done by
hand.

Also cleaned up a bit in the test suites so now more work.