Fix call to deprecated erlang:get_stacktrace/0
We define a macro ?CATCH/3 to handle try/catch exceptions which get a
stacktrace in the right way. Define the macro in lfe.hrl and move
wuote macros and ?IF there as well. We are using the ?CATCH macro
defined by Maxim in PR https://github.com/rvirding/lfe/pull/391 .

Co-authored-by: Robert Virding <rvirding@gmail.com>