For make-record and update-record make key/value list an a-list
We are going more lispy. This needed a lot of small changes in many
places which in many files "knew" about the structures.  Also allow
(field) when defining a record. We can also now handle records in
patterns using make-record and record-index.

We also took the chance to cleanup the code.