emacs: Add indentation rules for data-like syntax forms
Before:

    #m(a b
         c d)

After:

    #m(a b
       c d)

Same goes for '(), `(), #(), #m(), #b().

There's a slight downside with the default settings that in macros like this:

    (defmacro with-checked-caps body
      `(if (lobby:check-cap lobby event)
        ,@body)

As you can see, the `if` is not indented correctly, since this form is
treated as data, not as code.

The introduced behavior can be fully disabled by setting

    (setq lfe-indent-sequence-depth 0)

By default this variable is 1, settings the variable to higher values will
inspect more levels upwards for quoted forms or forms inside data structures.
Example for depth = 5:

    (defmacro with-checked-caps body
      `(if (lobby:check-cap lobby event)
        ,@body
        (progn
         (log-warn `#m(reason nocap event-type ,event-type
                       event ,event state ,state))
         (case event-type
          ((tuple 'call from)
           (tuple 'keep_state_and_data (list (tuple 'reply from #(error nocap)))))
          ('cast
           (tuple 'keep_state_and_data))))))